added dashboard using baseline model

+++ utils.sh
@@ -19,15 +19,26 @@ else
         cd 02-train
     elif [[ "$ACTION" == 'reset-explore' ]]; then
         cd 03-explore
+    elif [[ "$ACTION" == 'reset-upload' ]]; then
+        cd 04-upload
     elif [[ "$ACTION" == 'reset-dash' ]]; then
-        cd 04-dash
+        cd 05-dash
+    elif [[ "$ACTION" == 'reset-app' ]]; then
+        cd 06-app
     else
-        cd 05-cleanup
+        cd 07-cleanup
     fi
 
-    cd notebooks
-    rm -rf .ipynb_checkpoints/
+    if [[ "$ACTION" -ne 'reset-app' ]]; then
+        cd notebooks
+        rm -rf .ipynb_checkpoints/
 
-    cd ../../../src
+        cd ../../../src
+    fi
+
+    if [[ "$ACTION" == 'reset-app' ]]; then
+        cd src/
+    fi
+
     rm -rf __pycache__/
 fi
